test(segment): migrate a11y
